pyproject.toml
[project]
name = "docx2markdown"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "lxml>=6.0.2",
    "python-docx>=1.2.0",
]
